refactor: retire general KPN CMS store -- registry keeps datasets only
