+++ README.md
@@ -47,40 +47,21 @@ Docker will use environment variables in your .env file too so you can override 
 - Pushes/PR merges to main will deploy to QA via Github actions.
 - Publishing a tag pointing to a commit in the main branch will deploy that version to production
 
-## KPN CMS content (kp4cd.org replacement)
+## KPN dataset info (kp4cd.org dataset content)
 
-The `/api/kpn/*` endpoints serve the Drupal CMS content formerly hosted on kp4cd.org
-(news feeds, portal front content, help book, dataset info, etc.) in the exact
-response shapes the portal frontends expect. Rationale: kp4cd.org is being retired
-but is still the only live source of these views (hugeampkpncms.org serves a
-different API generation), so this store + import + compatible API replace it.
-Full audit and runbook live in the KP4CD Re-architecture project's support folder.
-
-`datasetinfo` is the one exception: it is now served from the registry's own
-`kp_datasets` table (the system of record for dataset content), populated and
+`GET /api/kpn/rest/views/datasetinfo` serves KP dataset info from the registry's
+own `kp_datasets` table — the system of record — in the Drupal-compatible
+response shape the portal frontends expect. The table is populated and
 refreshed by `python -m scripts.migrate_kp_datasets` (reads the kp4cd Drupal
-RDS directly via the `kp4cd-220214` secret). `scripts/import_kpn_cms.py` no
-longer touches `datasetinfo`.
+RDS directly via the `kp4cd-220214` secret; idempotent, re-runnable).
 
-Populate / refresh the content snapshot (requires kp4cd.org to be reachable):
+Embedded images in dataset bodies are mirrored to a dedicated public-content
+bucket, `KPN_CMS_ASSETS_BUCKET` (default `dig-kpn-cms-assets`), and served via
+`GET /api/kpn/files/{path}` (presigned S3 redirect). The bucket must exist
+before the first migration run; the API role needs read on it, the migration
+runner needs write.
 
-    python -m scripts.import_kpn_cms --dry-run   # review the URL set first
-    python -m scripts.import_kpn_cms             # import + mirror embedded assets to S3
-
-Config:
-- `KPN_CMS_PROXY_ON_MISS` (default `true`) — while kp4cd.org is alive, unknown
-  requests are proxied there, persisted, and logged to `cms_request_miss`.
-  Set to `false` at domain retirement.
-- `KPN_CMS_SOURCE_HOST` (default `https://kp4cd.org`)
-- Assets land in a dedicated public-content bucket, `KPN_CMS_ASSETS_BUCKET`
-  (default `dig-kpn-cms-assets`), under the `kpn-cms-assets/` prefix — separate
-  from the controlled-data registry bucket so its access policy can diverge
-  (e.g. public-read/CDN). The bucket must exist before the first asset import;
-  the API role needs read on it, the import runner needs write.
-
-Before retiring kp4cd.org, review the gap log — every request the store could not
-serve locally is recorded in the `cms_request_miss` table:
-
-    SELECT view_name, query_string, proxied, hit_count, last_seen
-    FROM cms_request_miss ORDER BY hit_count DESC;
+Non-dataset kp4cd.org content (news, portal front, help book, EGL methods,
+etc.) is out of scope for the registry and continues to be served from its
+existing source.
