--- tests/kpn_cms/fixtures/a2f_community_kps.json
@@ -1,1 +1,0 @@
-[{"title":"Skin Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/skinkp_symbol_only_2025.svg\u0022 alt=\u0022\u0022\u003E\u003C\/p\u003E","field_portal_address":"https:\/\/skin.hugeamp.org"},{"title":"Aging Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/images\/aging_small.svg\u0022 width=\u0022118\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"https:\/\/aging.hugeamp.org\/"},{"title":"Vision Genomics Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/images\/vision_genomics2.png\u0022 width=\u0022118\u0022\u003E\u003C\/p\u003E","field_portal_address":"https:\/\/hugeamp.org\/research.html?pageid=vision_landing"},{"title":"Kidney Disease Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/images\/kidney_small.svg\u0022 width=\u0022118\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"https:\/\/kidney.hugeamp.org\/"},{"title":"Autoimmune Disease Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/images\/autoimmune_small.svg\u0022 width=\u0022118\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"https:\/\/autoimmune.hugeamp.org\/"},{"title":"Ocular Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/images\/ocular_small.svg\u0022 width=\u0022118\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"https:\/\/ocular.hugeamp.org\/"},{"title":"Reproductive System Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/images\/reproductive_small.svg\u0022 width=\u0022118\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"https:\/\/reproductive.hugeamp.org\/"},{"title":"Sleep Disorder Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/sleep_small.svg\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"http:\/\/sleepdisordergenetics.org\/"},{"title":"Type 1 Diabetes Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/images\/t1d_small.svg\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"https:\/\/t1d.hugeamp.org\/"},{"title":"Type 2 Diabetes Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/t2d_small.svg\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"http:\/\/www.type2diabetesgenetics.org\/"},{"title":"Cardiovascular Disease Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/cvd_small.svg\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"http:\/\/www.broadcvdi.org\/"},{"title":"Cerebrovascular Disease Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/cd_small.svg\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"http:\/\/www.cerebrovascularportal.org\/"},{"title":"Common Metabolic Diseases Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/cmdkp_symbol_only_2024.svg\u0022 alt=\u0022\u0022\u003E\u003C\/p\u003E","field_portal_address":"https:\/\/hugeamp.org"},{"title":"Lipid Droplet Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/images\/lipid_small.svg\u0022 width=\u0022120\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"http:\/\/lipiddroplet.org\/"},{"title":"Lung Disease Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/images\/lung_small.svg\u0022 width=\u0022118\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"https:\/\/lung.hugeamp.org\/"},{"title":"Musculoskeletal Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/msk.svg\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"http:\/\/mskkp.org"},{"title":"Neurodegenerative Disease Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/images\/ndkp_small.svg\u0022 width=\u0022120\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"https:\/\/ndkp.hugeamp.org\/"},{"title":"Non-Additive Genetic Effects Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/images\/nage_small.svg\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"https:\/\/nage.hugeamp.org\/"},{"title":"Collaborate on methods","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/images\/collaborate_methods_icon.svg\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"https:\/\/kp4cd.org\/collaborate-methods"},{"title":"Collaborate to create a new Knowledge Portal","body":"\u003Cp\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/images\/create_kp_icon.svg\u0022 \/\u003E\u003C\/p\u003E\r\n","field_portal_address":"https:\/\/kp4cd.org\/collaborate-KP"}]
