--- tests/kpn_cms/fixtures/content_by_id_sample.json
@@ -1,1 +1,0 @@
-[{"title":"August 6, 2026 KPN Webinar Recording Now LIVE","body":"\u003Cp\u003E\ud83e\uddec dkNET Computational Core \u2014 A Cloud-native, Collaborative AI\/ML Platform for Biomedical Research\u003C\/p\u003E\u003Cp\u003EThe dkNET Computational Core (dknet-ai.org) provides a cloud-native, collaborative platform for biomedical data science, artificial intelligence, machine learning, and literature intelligence. Powered by the open-source Apache Texera (Incubating) project, it enables researchers to create reproducible, browser-accessible analyses; work together through simultaneous workflow editing and execution; and use large language model agents for no-code or low-code bioinformatics analysis and guided exploration of the medical literature.\u003C\/p\u003E\u003Cp\u003E\u003Cbr \/\u003EThis talk presents the platform\u2019s researcher-facing capabilities and underlying technology architecture. The system combines a browser-based graphical interface, drag-and-drop workflows, live execution visualization, natural-language assistance, rich AI and machine-learning operators, interactive debugging, and extensibility through user-defined functions in Python, R, Java, and Scala. Its cloud-native execution layer uses containerized and cluster-based infrastructure to support scalable, shared, and reproducible research while keeping researchers actively involved in the analysis process. Together, these capabilities lower barriers to advanced computational methods, support reproducible team science, and provide scalable shared infrastructure for bioinformaticians and research communities.\u003C\/p\u003E\u003Cp\u003E\u003Cbr \/\u003E\ud83c\udf99\ufe0f Featured speaker: Prof. Chen Li, Department of Computer Science, UC Irvine\u003C\/p\u003E\u003Cp\u003EProf. Chen Li is a professor in the Department of Computer Science at UC Irvine. He received his Ph.D. degree in Computer Science from Stanford University, and his M.S. and B.S. in Computer Science from Tsinghua University, China. \u00a0His research interests are in the fields of data management, data science, AI\/ML, databases, data-intensive computing, search, and visualization. \u00a0He was a co-founder and CTO of a startup to commercialize his research. He was a recipient of an NSF CAREER award, an ACM Distinguished Member, and an IEEE fellow.\u003C\/p\u003E\u003Cp\u003E\ud83d\udcbb Watch the recording: https:\/\/youtu.be\/J15LylFMTJA\u003C\/p\u003E\u003Cimg src=\u0022https:\/\/kp4cd.org\/sites\/default\/files\/News%20Items\/August%206%2C%202026%20KPN%20Webinar.png\u0022 data-entity-uuid=\u0022\u0022 data-entity-type=\u0022file\u0022 alt=\u0022\u0022 width=\u0022960\u0022 height=\u0022540\u0022 loading=\u0022lazy\u0022 \/\u003E"}]
