--- tests/kpn_cms/fixtures/eglmethod_sample.json
@@ -1,1 +1,0 @@
-[{"title":"EGL method - CARDIoGRAM","body":"\u003Cp\u003ETo prioritize likely causal genes for each of 241 genomewide-significant CAD association signals, eight locus-based or similarity-based predictors of causal genes were considered:\u003C\/p\u003E\n\n\u003Col\u003E\u003Cli\u003EGenes harboring pathogenic (or likely pathogenic) variants responsible for monogenic disorders of cardiovascular relevance according to \u003Ca href=\u0022https:\/\/www.ncbi.nlm.nih.gov\/clinvar\/\u0022 target=\u0022_blank\u0022\u003EClinVar\u003C\/a\u003E\u003C\/li\u003E\n\t\u003Cli\u003EGenes encoding proteins of established causal relevance to CAD per Mendelian randomization studies, or that are targets for established cardiovascular drugs;\u003C\/li\u003E\n\t\u003Cli\u003EThe two genes with the highest Polygenic Priority Score (\u003Ca href=\u0022https:\/\/www.medrxiv.org\/content\/10.1101\/2020.09.08.20190561v1\u0022 target=\u0022_blank\u0022\u003EPoPS\u003C\/a\u003E) within 500kb of the sentinel variant;\u003C\/li\u003E\n\t\u003Cli\u003EGenes with eQTLs in CAD-relevant tissues from \u003Ca href=\u0022https:\/\/science.sciencemag.org\/content\/353\/6301\/827.long\u0022 target=\u0022_blank\u0022\u003ESTARNET\u003C\/a\u003E or \u003Ca href=\u0022https:\/\/www.gtexportal.org\/home\/\u0022 target=\u0022_blank\u0022\u003EGTEx\u003C\/a\u003E\u003C\/li\u003E\n\t\u003Cli\u003EGenes that display cardiovascular-relevant phenotypes in knock-out mice from the \u003Ca href=\u0022https:\/\/www.mousephenotype.org\/\u0022 target=\u0022_blank\u0022\u003EInternational Mouse Phenotyping Consortium\u003C\/a\u003E or the \u003Ca href=\u0022http:\/\/www.informatics.jax.org\/\u0022 target=\u0022_blank\u0022\u003EMouse Genome Informatics database\u003C\/a\u003E\u003C\/li\u003E\n\t\u003Cli\u003EGenes containing protein-altering (i.e. missense or predicted loss-of-function) variants that are in strong LD (r\u003Csup\u003E2\u003C\/sup\u003E\u22650.8) with the CAD sentinel variant;\u003C\/li\u003E\n\t\u003Cli\u003EGenes containing rare coding variants that have been associated with CAD risk in previous whole-exome sequencing or array-based studies;\u003C\/li\u003E\n\t\u003Cli\u003EThe nearest gene to the CAD sentinel variant.\u003C\/li\u003E\n\u003C\/ol\u003E\u003Cp\u003ESee the figure below for a summary. Genes were prioritized as \u0022likely causal\u0022 for each sentinel association using a consensus-based approach, selecting the gene with the highest, unweighted sum (0-8) of evidence across all eight predictors. Most likely causal genes were assigned for 206 associations where there were at least two concordant predictors for a gene, prioritizing 185 distinct genes. 94 of these genes had three or more concordant predictors, so were considered \u0022strongly prioritized\u0022.\u003C\/p\u003E\n\n\u003Cp\u003EFurther details can be found in \u003Ca href=\u0022https:\/\/www.medrxiv.org\/content\/10.1101\/2021.05.24.21257377v1\u0022 target=\u0022_blank\u0022\u003EAragam \u003Ci\u003Eet al., medRxiv, \u003C\/i\u003E2021\u003C\/a\u003E. \u003Ca href=\u0022https:\/\/www.youtube.com\/watch?v=AB43ljZ3Fxk\u0022 target=\u0022_blank\u0022\u003EWatch an \u0022explainer\u0022 video\u003C\/a\u003E about this study, created by Jeanette Erdmann.\u003C\/p\u003E\n\n\u003Cp style=\u0022text-align: center\u0022\u003E\u003Cimg align=\u0022center\u0022 alt=\u0022\u0022 src=\u0022http:\/\/kp4cd.org\/sites\/default\/files\/images\/Aragam_egl_method.png\u0022 style=\u0022width:800px;\u0022 \/\u003E\u003C\/p\u003E\n","field_research_from":"cardiogram"}]
