--- tests/kpn_cms/fixtures/help_book.json
@@ -1,1 +1,0 @@
-[{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Data","nid_2":"912","title":"Data","nid":"912"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Data","nid_2":"912","title":"Understanding genetic association data","nid":"1036"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Data","nid_2":"912","title":"Genetic association data","nid":"935"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Data","nid_2":"912","title":"Credible sets","nid":"940"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Data","nid_2":"912","title":"Genomic annotations","nid":"936"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Data","nid_2":"912","title":"Curated knowledge","nid":"938"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Data","nid_2":"912","title":"Contributing data","nid":"939"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Data","nid_2":"912","title":"Download files","nid":"937"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Methods","nid_2":"908","title":"Bottom-line integrative analysis","nid":"911"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Methods","nid_2":"908","title":"LD clumping","nid":"909"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Methods","nid_2":"908","title":"Credible Sets to Cell Type (CS2CT)","nid":"1517"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Methods","nid_2":"908","title":"Methods","nid":"908"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Methods","nid_2":"908","title":"MAGMA","nid":"955"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Methods","nid_2":"908","title":"LDSC and S-LDSC","nid":"1094"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Methods","nid_2":"908","title":"Filtering variants with masks","nid":"1043"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Pages and tools","nid_2":"914","title":"Gene page","nid":"947"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Pages and tools","nid_2":"914","title":"Region page","nid":"946"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Pages and tools","nid_2":"914","title":"Variant page","nid":"948"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Pages and tools","nid_2":"914","title":"Pages and tools","nid":"914"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Pages and tools","nid_2":"914","title":"Phenotype page","nid":"949"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Pages and tools","nid_2":"914","title":"Tissue page","nid":"1441"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Pages and tools","nid_2":"914","title":"Gene Sifter","nid":"956"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Pages and tools","nid_2":"914","title":"Genetic Association Interactive Tool (GAIT)","nid":"959"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Pages and tools","nid_2":"914","title":"Human Genetic Evidence (HuGE) Calculator","nid":"961"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Pages and tools","nid_2":"914","title":"Predicted effector genes","nid":"957"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Pages and tools","nid_2":"914","title":"Signal Sifter","nid":"958"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Pages and tools","nid_2":"914","title":"Programmatic access","nid":"915"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Workflows","nid_2":"950","title":"Genetic support workflow","nid":"951"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Workflows","nid_2":"950","title":"Multi-trait analysis","nid":"952"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Workflows","nid_2":"950","title":"Integrative analysis","nid":"953"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Workflows","nid_2":"950","title":"Rare variant gene-level associations workflow","nid":"954"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Workflows","nid_2":"950","title":"Workflows","nid":"950"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"FAQ","nid_2":"934","title":"FAQ","nid":"934"},{"title_1":"Knowledge Portal Guide","nid_1":"786","title_2":"Videos, webinars, and presentations","nid_2":"1039","title":"Videos, webinars, and presentations","nid":"1039"}]
