--- tests/kpn_cms/fixtures/news2vueportal_md.json
@@ -1,1 +1,0 @@
-[{"title":"August 6, 2026 KPN Webinar Recording Now LIVE","body":"Prof. Chen Li from the Department of Computer Science, UC Irvine presented: \u0022dkNET Computational Core \u2014 A Cloud-native, Collaborative AI\/ML Platform for Biomedical Research.\u0022","view_node":"\u003Ca href=\u0022\/node\/1770\u0022 hreflang=\u0022en\u0022\u003Eview\u003C\/a\u003E","nid":"1770","field_portals":"md, cvd, cd, sleep, t1d, t2d, lung, nage, msk, a2f, kidney, aging","field_image":"https:\/\/kp4cd.org\/sites\/default\/files\/news_thumbnails\/August%206%2C%202026%20KPN%20Webinar_0.png"},{"title":"July 2026 CFDE Webinar Recording Now LIVE","body":"Unable to join today\u0027s CFDE Public Webinar Series featuring the Human Reference Atlas?","view_node":"\u003Ca href=\u0022\/node\/1769\u0022 hreflang=\u0022en\u0022\u003Eview\u003C\/a\u003E","nid":"1769","field_portals":"md, cvd, cd, sleep, t1d, t2d, lung, nage, msk, a2f, kidney, aging","field_image":"https:\/\/kp4cd.org\/sites\/default\/files\/news_thumbnails\/CFDE%20Public%20Webinar%20Series%20-%20Friday%2C%20July%2030%2C%202026%20%282%29.png"},{"title":"July 23, 2026 KPN Webinar Recording Now LIVE","body":"Unable to join today\u0027s KPN webinar? The recording is now live.","view_node":"\u003Ca href=\u0022\/node\/1768\u0022 hreflang=\u0022en\u0022\u003Eview\u003C\/a\u003E","nid":"1768","field_portals":"md, cvd, cd, sleep, t1d, t2d, lung, nage, msk, a2f, kidney, aging","field_image":"https:\/\/kp4cd.org\/sites\/default\/files\/news_thumbnails\/July%2023%2C%202026%20KPN%20Webinar_0.png"},{"title":"Join us for the August 6, 2026 KPN Webinar!","body":"Prof. Chen Li from the Department of Computer Science, UC Irvine will present: \u0022dkNET Computational Core \u2014 A Cloud-native, Collaborative AI\/ML Platform for Biomedical Research.\u0022","view_node":"\u003Ca href=\u0022\/node\/1767\u0022 hreflang=\u0022en\u0022\u003Eview\u003C\/a\u003E","nid":"1767","field_portals":"md, cvd, cd, sleep, t1d, t2d, lung, nage, msk, a2f, kidney, aging","field_image":"https:\/\/kp4cd.org\/sites\/default\/files\/news_thumbnails\/August%206%2C%202026%20KPN%20Webinar.png"},{"title":"Join us for the July 23rd, 2026 KPN Webinar!","body":"Trang Nguyen will present on \u0022Rare-variant association studies: a primer, a problem, and a proposal for baseline masks.\u0022","view_node":"\u003Ca href=\u0022\/node\/1766\u0022 hreflang=\u0022en\u0022\u003Eview\u003C\/a\u003E","nid":"1766","field_portals":"md, cvd, cd, sleep, t1d, t2d, lung, nage, msk, a2f, kidney, aging","field_image":"https:\/\/kp4cd.org\/sites\/default\/files\/news_thumbnails\/July%2023%2C%202026%20KPN%20Webinar.png"},{"title":"Join us on Friday, July 30th, 2026 for our next session of the Public Webinar Series!","body":"Hear about the construction and usage of the Human Reference Atlas developed in HuBMAP \u0026 SenNet.","view_node":"\u003Ca href=\u0022\/node\/1765\u0022 hreflang=\u0022en\u0022\u003Eview\u003C\/a\u003E","nid":"1765","field_portals":"md, cvd, cd, sleep, t1d, t2d, lung, nage, msk, a2f, kidney, aging","field_image":"https:\/\/kp4cd.org\/sites\/default\/files\/news_thumbnails\/CFDE%20Public%20Webinar%20Series%20-%20Friday%2C%20July%2030%2C%202026.png"},{"title":"Introducing PEGASUS, a new community framework for Effector Gene reporting!","body":"A new community framework for Effector Gene reporting, PEGASUS, is available from the Knowledge Portal Network and GWAS Catalog, in support from OpenTargets and the scientific community! ","view_node":"\u003Ca href=\u0022\/node\/1764\u0022 hreflang=\u0022en\u0022\u003Eview\u003C\/a\u003E","nid":"1764","field_portals":"md, cvd, cd, sleep, t1d, t2d, lung, nage, msk, a2f, kidney, aging","field_image":"https:\/\/kp4cd.org\/sites\/default\/files\/news_thumbnails\/pegasus_icon.png"},{"title":"Recording now available for the July 9, 2026 PanKbase Webinar!","body":"Unable to join today\u0027s PanKbase webinar? The recording is now live.","view_node":"\u003Ca href=\u0022\/node\/1763\u0022 hreflang=\u0022en\u0022\u003Eview\u003C\/a\u003E","nid":"1763","field_portals":"md, cvd, cd, sleep, t1d, t2d, lung, nage, msk, a2f, kidney, aging","field_image":"https:\/\/kp4cd.org\/sites\/default\/files\/news_thumbnails\/July%209th%2C%202026%20KPN%20Webinar_0.png"},{"title":"Join us for the July 9th, 2026 PanKbase Office Hours","body":"Kyle Gaulton, PhD, UCSD will present Identifying disease-associated cellular states in the pancreas","view_node":"\u003Ca href=\u0022\/node\/1762\u0022 hreflang=\u0022en\u0022\u003Eview\u003C\/a\u003E","nid":"1762","field_portals":"md, cvd, cd, sleep, t1d, t2d, lung, nage, msk, a2f, kidney, aging","field_image":"https:\/\/kp4cd.org\/sites\/default\/files\/news_thumbnails\/July%209th%2C%202026%20KPN%20Webinar.png"},{"title":"June 2026 CFDE Webinar Recording Now LIVE","body":"Unable to join today\u0027s CFDE Public Webinar Series featuring Cell Maps for AI (CM4AI)?","view_node":"\u003Ca href=\u0022\/node\/1761\u0022 hreflang=\u0022en\u0022\u003Eview\u003C\/a\u003E","nid":"1761","field_portals":"md, cvd, cd, sleep, t1d, t2d, lung, nage, msk, a2f, kidney, aging","field_image":"https:\/\/kp4cd.org\/sites\/default\/files\/news_thumbnails\/CFDE%20Public%20Webinar%20Series%20-%20Friday%2C%20June%2026%2C%202026_0.png"}]
