--- tests/kpn_cms/fixtures/paperheadermenu_sample.json
@@ -1,1 +1,0 @@
-[{"title":"APOL1 header menu","body":"\u003Cul\u003E\r\n\t\u003Cli\u003E\u003Ca href=\u0022\/paper.html?paper=apol1_portal\u0022\u003EHome\u003C\/a\u003E\u003C\/li\u003E\r\n\t\u003Cli\u003E\u003Ca href=\u0022\/paper.html?paper=apol1_about\u0022\u003EAbout\u003C\/a\u003E\u003C\/li\u003E\r\n\t\u003Cli\u003E\u003Ca href=\u0022\/paper.html?paper=apol1_tutorial\u0022\u003ETutorial\u003C\/a\u003E\u003C\/li\u003E\r\n\t\u003Cli\u003E\u003Ca href=\u0022\/paper.html?paper=apol1_downloads\u0022\u003EDownloads\u003C\/a\u003E\u003C\/li\u003E\r\n\u003C\/ul\u003E\r\n","field_paper":"apol1_portal, apol1_shiny_app, apol1_gene_view, apol1_diff_exp, apol1_diff_co_exp, apol1_cor, apol1_about, apol1_downloads, apol1_tutorial, apol1_david"}]
