+++ tests/kpn_cms/test_migrate_kp_datasets.py
@@ -7,7 +7,6 @@ from sqlalchemy import text
 
 from dataregistry.api.db import DataRegistryReadWriteDB
 from dataregistry.api import kp_datasets_query as kq
-from dataregistry.api import kpn_cms_query as q
 from scripts import migrate_kp_datasets as mig
 
 engine = DataRegistryReadWriteDB().get_engine()
@@ -16,7 +15,6 @@ engine = DataRegistryReadWriteDB().get_engine()
 def _clean():
     with engine.connect() as con:
         con.execute(text("TRUNCATE TABLE kp_datasets"))
-        con.execute(text("TRUNCATE TABLE cms_content_item"))
         con.execute(text("TRUNCATE TABLE cms_asset"))
         con.commit()
 
@@ -73,18 +71,13 @@ def test_node_to_row_converts_types_and_defaults():
 
 
 @mock_aws
-def test_run_migration_upserts_links_and_cleans_cms(monkeypatch):
+def test_run_migration_upserts_and_links_datasets(monkeypatch):
     _clean()
     monkeypatch.setattr(mig, 'fetch_drupal_nodes',
                         lambda: [_node(1, 'MIGTEST_A'), _node(2, 'MIGTEST_B', status=0)])
-    q.replace_view_rows(engine, 'datasetinfo', 'item_key', 'old',
-                        [{'view_name': 'datasetinfo', 'portal': None, 'nid': None,
-                          'item_key': 'old', 'payload': '{}', 'search_text': None,
-                          'sort_order': 0}])
     s3 = boto3.client('s3', region_name='us-east-1'); s3.create_bucket(Bucket='kpn-cms-test')
     report = mig.run_migration(engine, s3, 'kpn-cms-test', skip_assets=True)
     assert report['nodes'] == 2 and report['published'] == 1
-    assert report['cms_rows_deleted'] == 1
     assert kq.get_by_dataset_id(engine, 'MIGTEST_A')['title'] == 't1'
     # idempotent: run again, same row count
     mig.run_migration(engine, s3, 'kpn-cms-test', skip_assets=True)
